% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Connect.R
\name{tt2.QF.Connect}
\alias{tt2.QF.Connect}
\title{Connects to a TT server with QuoteFeed client}
\usage{
tt2.QF.Connect(address = "", login = "", password = "")
}
\arguments{
\item{address}{Url of the server}

\item{login}{account you login}

\item{password}{password of the account}
}
\description{
Connects to a TT server with QuoteFeed client
}
