% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Connect.R
\name{tt2.QF.Disconnect}
\alias{tt2.QF.Disconnect}
\title{Disconnect from a TT server with QuoteFeed client}
\usage{
tt2.QF.Disconnect()
}
\description{
Disconnect from a TT server with QuoteFeed client
}
\examples{
tt2.QF.Disconnect()

}
