% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Currencies.R
\name{tt2.QF.GetCurrencyList}
\alias{tt2.QF.GetCurrencyList}
\title{Gets the currencies as requested}
\usage{
tt2.QF.GetCurrencyList()
}
\description{
Gets the currencies as requested
}
\examples{
tt2.QF.GetCurrencyList()

}
