% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Symbols.R
\name{tt2.QF.GetSymbolHiddenLimitOrderMarginReduction}
\alias{tt2.QF.GetSymbolHiddenLimitOrderMarginReduction}
\title{Get Symbol field}
\usage{
tt2.QF.GetSymbolHiddenLimitOrderMarginReduction()
}
\description{
Get Symbol field
}
