% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Connect.R
\name{tt2.QS.Disconnect}
\alias{tt2.QS.Disconnect}
\title{Disconnect from a TT server with QuoteStore client}
\usage{
tt2.QS.Disconnect()
}
\description{
Disconnect from a TT server with QuoteStore client
}
\examples{
tt2.QS.Disconnect()

}
