% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bars.R
\name{tt2.QS.DownloadBars}
\alias{tt2.QS.DownloadBars}
\title{Gets the bars as requested}
\usage{
tt2.QS.DownloadBars(symbol, priceType, periodicity, from, to)
}
\arguments{
\item{symbol}{Symbol}

\item{priceType}{PriceType}

\item{periodicity}{Periodicity}

\item{from}{From}

\item{to}{To}
}
\description{
Gets the bars as requested
}
