% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ticks.R
\name{tt2.QS.DownloadTicks}
\alias{tt2.QS.DownloadTicks}
\title{Gets the ticks as requested}
\usage{
tt2.QS.DownloadTicks(symbol, from, to)
}
\arguments{
\item{symbol}{Symbol}

\item{from}{From}

\item{to}{to}
}
\description{
Gets the ticks as requested
}
