% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TickL2.R
\name{tt2.QS.DownloadTicksL2}
\alias{tt2.QS.DownloadTicksL2}
\title{Gets the ticks L2 as requested}
\usage{
tt2.QS.DownloadTicksL2(symbol, from, to)
}
\arguments{
\item{symbol}{Symbol}

\item{from}{From}

\item{to}{To}
}
\description{
Gets the ticks L2 as requested
}
