% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bars.R
\name{tt2.QS.GetBarList}
\alias{tt2.QS.GetBarList}
\title{Gets the bars as requested}
\usage{
tt2.QS.GetBarList(symbol, priceType, periodicity, from, count)
}
\arguments{
\item{symbol}{Symbol}

\item{priceType}{PriceType}

\item{periodicity}{Periodicity}

\item{from}{From}

\item{count}{Count}
}
\description{
Gets the bars as requested
}
