% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TickL2.R
\name{tt2.QS.GetTickL2List}
\alias{tt2.QS.GetTickL2List}
\title{Gets the ticks L2 as requested}
\usage{
tt2.QS.GetTickL2List(symbol, from, count)
}
\arguments{
\item{symbol}{Symbol}

\item{from}{From}

\item{count}{Count}
}
\description{
Gets the ticks L2 as requested
}
