% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ticks.R
\name{tt2.QS.GetTickList}
\alias{tt2.QS.GetTickList}
\title{Gets the ticks as requested}
\usage{
tt2.QS.GetTickList(symbol, from, count)
}
\arguments{
\item{symbol}{Symbol}

\item{from}{From}

\item{count}{Count}
}
\description{
Gets the ticks as requested
}
